requirements.txt
requests
python-dotenv
flask
flask-sqlalchemy
flask-migrate 
flask-bcrypt
flask-login
markdown